revamping loading

+++ src/AssuredLibrary.sol
@@ -2,15 +2,14 @@
 pragma solidity ^0.8.26;
 
 library AssuredLibrary {
-    
     struct Inspectors {
         address inspector;
         bool valid;
         uint8 assetType;
-        uint assetInspected;
+        uint256 assetInspected;
         bool currentlyInspecting;
         bool status;
-        uint inspectorId;
+        uint256 inspectorId;
     }
 
     enum InspectionStatus {
@@ -21,76 +20,73 @@ library AssuredLibrary {
     }
 
     struct Property {
-        uint id;
+        uint256 id;
         address owner;
         address inspector;
         InspectionStatus status;
         string assetType;
-        uint assetLocation;
-        uint categoryofAsset;
-        uint ageOfAsset;
-        uint safetyFeatures;
-        uint propertyValue;
-        uint premium;
+        uint256 assetLocation;
+        uint256 categoryofAsset;
+        uint256 ageOfAsset;
+        uint256 safetyFeatures;
+        uint256 propertyValue;
+        uint256 premium;
         string addr;
         bool paid;
     }
 
     struct Vehicle {
-        uint id;
+        uint256 id;
         address owner;
         address inspector;
         InspectionStatus status;
-        uint safetyFeatures;
-        uint vehicleValue;
-        uint vehicleAge;
-        uint premium;
-        uint vehicleType;
-        uint claimHistory;
-        uint driverAge;
+        uint256 safetyFeatures;
+        uint256 vehicleValue;
+        uint256 vehicleAge;
+        uint256 premium;
+        uint256 vehicleType;
+        uint256 claimHistory;
+        uint256 driverAge;
         string ownersAddress;
         bool paid;
     }
 
     function calculatePropertyInsurancePremium(
-        uint location,
-        uint propertyType,
-        uint age,
-        uint protections,
-        uint propertyValue
-    ) internal pure returns (uint) {
+        uint256 location,
+        uint256 propertyType,
+        uint256 age,
+        uint256 protections,
+        uint256 propertyValue
+    ) internal pure returns (uint256) {
         // Base rate for scaling
-        uint baseRate = 1;
+        uint256 baseRate = 1;
 
         // Combine risk factors into a single step
-        uint riskFactors = location + propertyType - protections + age;
+        uint256 riskFactors = location + propertyType - protections + age;
 
         // Ensure propertyValue is valid and use it to scale the premium
         require(propertyValue > 0, "Property value must be greater than zero");
 
         // Calculate the premium in a single step, using a more appropriate scaling factor
-        uint premium = (baseRate * riskFactors * propertyValue) / 10000;
+        uint256 premium = (baseRate * riskFactors * propertyValue) / 10000;
 
         return premium;
     }
 
     function calculateVehicleInsurancePremium(
-        uint vehicleType, // Type of vehicle
-        uint vehicleAge, // Age of the vehicle in years
-        uint driverAge, // Age of the driver
-        uint safetyFeatures, // Number of safety features (a score / 100 determined by inspectors)
-        uint vehicleValue, // Value of the vehicle
-        uint claimHistory // Number of claims the driver has had
-    ) internal pure returns (uint) {
+        uint256 vehicleType, // Type of vehicle
+        uint256 vehicleAge, // Age of the vehicle in years
+        uint256 driverAge, // Age of the driver
+        uint256 safetyFeatures, // Number of safety features (a score / 100 determined by inspectors)
+        uint256 vehicleValue, // Value of the vehicle
+        uint256 claimHistory // Number of claims the driver has had
+    ) internal pure returns (uint256) {
         // Base rate for scaling, and safety feature discount
-        uint baseRate = 2;
+        uint256 baseRate = 2;
 
         // Risk factor calculation in one step
-        uint riskFactors = (vehicleType * 10) +
-            (vehicleAge * 5) +
-            ((driverAge < 25 || driverAge > 70) ? 20 : 10) -
-            (safetyFeatures * 3) +
-            (claimHistory * 15);
+        uint256 riskFactors = (vehicleType * 10) + (vehicleAge * 5) + ((driverAge < 25 || driverAge > 70) ? 20 : 10)
+            - (safetyFeatures * 3) + (claimHistory * 15);
 
         // Ensure vehicle value is valid to avoid division by zero
         require(vehicleValue > 0, "Vehicle value must be greater than zero");
