+++ src/CentralStorage.sol
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.26;
+
 import "./AssuredLibrary.sol";
 
 contract CentralStorage {
@@ -7,19 +8,19 @@ contract CentralStorage {
     using AssuredLibrary for AssuredLibrary.Property;
     using AssuredLibrary for AssuredLibrary.InspectionStatus;
 
-    uint public propertyId;
-    uint public vehicleId;
+    uint256 public propertyId;
+    uint256 public vehicleId;
 
-    uint inspectorsId;
+    uint256 inspectorsId;
     address public daoAddress;
 
     AssuredLibrary.Property[] public pendingProperties;
     AssuredLibrary.Vehicle[] public pendingVehicles;
 
-    mapping(uint => AssuredLibrary.Property) propertyIds;
-    mapping(uint => AssuredLibrary.Vehicle) vehicleIds;
+    mapping(uint256 => AssuredLibrary.Property) propertyIds;
+    mapping(uint256 => AssuredLibrary.Vehicle) vehicleIds;
     mapping(address => AssuredLibrary.Inspectors) public inspectors;
-    mapping(uint => uint) public data; // Shared storage (can be any data structure)
+    mapping(uint256 => uint256) public data; // Shared storage (can be any data structure)
 
     constructor(address _daoAddress) {
         daoAddress = _daoAddress;
@@ -28,23 +29,18 @@ contract CentralStorage {
     function onlyDao() public view returns (address, address) {
         // require(msg.sender == daoAddress, "Only Owner can Perform this Action");
         // return (tx.origin, msg.sender);
-        return(msg.sender, msg.sender);
+        return (msg.sender, msg.sender);
     }
 
     // Function to set data
-    function setInspector(
-        address _address,
-        AssuredLibrary.Inspectors memory _inspector
-    ) public {
+    function setInspector(address _address, AssuredLibrary.Inspectors memory _inspector) public {
         inspectors[_address] = _inspector;
         AssuredLibrary.Inspectors storage newInspector = inspectors[_address];
         newInspector.inspector = _address;
     }
 
     // Function to get data
-    function getInspector(
-        address _adddress
-    ) public view returns (AssuredLibrary.Inspectors memory _inspector) {
+    function getInspector(address _adddress) public view returns (AssuredLibrary.Inspectors memory _inspector) {
         return inspectors[_adddress];
     }
 
@@ -56,7 +52,7 @@ contract CentralStorage {
         inspectorsId++;
     }
 
-    function returnInspectorsId() public view returns (uint) {
+    function returnInspectorsId() public view returns (uint256) {
         return inspectorsId;
     }
 
@@ -64,7 +60,7 @@ contract CentralStorage {
         propertyId++;
     }
 
-    function returnPropertyId() public view returns (uint) {
+    function returnPropertyId() public view returns (uint256) {
         return propertyId;
     }
 
@@ -72,9 +68,7 @@ contract CentralStorage {
         propertyIds[propertyId] = _property;
     }
 
-    function getProperty(
-        uint _propertyId
-    ) public view returns (AssuredLibrary.Property memory _inspector) {
+    function getProperty(uint256 _propertyId) public view returns (AssuredLibrary.Property memory _inspector) {
         return propertyIds[_propertyId];
     }
 
@@ -82,7 +76,7 @@ contract CentralStorage {
         vehicleId++;
     }
 
-    function returnVehicleId() public view returns (uint) {
+    function returnVehicleId() public view returns (uint256) {
         return vehicleId;
     }
 
@@ -90,9 +84,7 @@ contract CentralStorage {
         vehicleIds[vehicleId] = _vehicle;
     }
 
-    function getVehicle(
-        uint _vehicleId
-    ) public view returns (AssuredLibrary.Vehicle memory _vehicle) {
+    function getVehicle(uint256 _vehicleId) public view returns (AssuredLibrary.Vehicle memory _vehicle) {
         return vehicleIds[_vehicleId];
     }
